# config/openapi_v2.yml
openapi: 3.0.3
info:
  title: ARASAAC API
  description: Backend API example API declaration
  termsOfService: http://swagger.io/terms/
  contact:
    email: juandacorreo@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
...
servers:
  # - url: https://api.arasaac.org/api/v2
  - url: http://localhost:3000/v2
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      x-eov-operation-handler: users
      description: Obtain user profile
      tags:
        - users
      security:
        - oAuth2:
            - admin
            - translator
            - read
            - write
      responses:
        200:
          description: User data like username, email, lists of favorites pictograms and materials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /materials/{id}:
    get:
      operationId: getMaterialById
      x-eov-operation-handler: materials
      description: Obtain a material based on its key.
      security:
        - oAuth2:
            - admin
            - translator
            - read
            - write
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The material id

      tags:
        - materials
      # define the type of response for Success "200" and Error
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMaterialResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /materials/days/{days}:
    get:
      operationId: getNewMaterials
      x-eov-operation-handler: materials
      description: Obtain materials published in the last number of days.
      security:
        - oAuth2:
            - admin
            - translator
            - read
            - write
      parameters:
        - in: path
          name: days
          required: true
          schema:
            type: integer
            default: 30
            minimum: 0
            maximum: 50
          description: Date range for materials published from the current date backwards. Minimum 0 days, maximum 50 days.
      tags:
        - materials
      # define the type of response for Success "200" and Error

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Materials'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /materials/new/{numItems}:
    get:
      operationId: getLastMaterials
      x-eov-operation-handler: materials
      description: Obtain last modified or published materials.
      security:
        - oAuth2:
            - admin
            - translator
            - read
            - write
      parameters:
        - in: path
          name: numItems
          required: true
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
          description: Total number of materials requested. Minimum 1, maximum 100.
      tags:
        - materials
      # define the type of response for Success "200" and Error

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Materials'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /materials/{language}/{searchText}:
    get:
      operationId: searchMaterials
      x-eov-operation-handler: materials
      description: Obtain a list of materials based on the search text parameter. It will search in the title and description fields. Language is used for search optimization.
      security:
        - oAuth2:
            - admin
            - translator
            - read
            - write
      tags:
        - materials
      # define the type of response for Success "200" and Error
      parameters:
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/SearchText'

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Materials'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    Language:
      name: language
      description: Languages than we currently use for pictograms searches.
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Locales'
...
    SearchText:
      name: searchText
      in: path
      description: Keywords for searches. By default, it makes an OR search on all the specified keywords. In case you would like to perform an exact phrase search (logical AND), you can do so by specifying double quotes in the search text. Prefixing a search keyword with – (minus sign) excludes all the documents that contain the negated term.
      required: true
      schema:
        type: string
...
  securitySchemes:
    oAuth2:
      type: oauth2
      description: See https://arasaac.org/developers
      flows:
        clientCredentials:
          tokenUrl: https://auth.arasaac.org/oauth/token/
          # scope read is provided by default
          scopes:
            admin: Manage users and material publication
            translator: add translations to materials
            read: search pictograms, materials, keywords
            write: upload materials
  schemas:
    Locales:
      type: string
      enum:
        - an
        - ar
        - bg
        - br
        - ca
        - de
        - el
        - en
        - es
        - et
        - eu
        - fa
        - fr
        - gl
        - he
        - hr
        - hu
        - it
        - ko
        - lt
        - lv
        - mk
        - nl
        - pl
        - pt
        - ro
        - ru
        - sk
        - sq
        - sv
        - sr
        - val
        - uk
        - zh
    ErrorResponse:
      required:
        - message
      properties:
        message:
          type: string
        err:
          type: string
...
    UserProfileResponse:
      properties:
        email:
          type: string
        name:
          type: string
        username:
          type: string
        lastlogin:
          type: string
        locale:
          $ref: '#/components/schemas/Locales'
        role:
          type: string
          enum:
            - admin
            - translator
            - user
        materials:
          type: object
        pictos:
          type: object
        provider:
          type: object
    Material:
      required:
        - authors
        - desc
        - files
        - title
      type: object
      properties:
        _id:
          type: string
        activity:
          type: array
          items:
            type: integer
        area:
          type: array
          items:
            type: integer
        authors:
          type: array
          items:
            type: object
        created:
          type: string
          format: date-time
        desc:
          type: string
        downloads:
          type: number
        file:
          type: object
        files:
          type: object
        id:
          type: integer
        screenshots:
          type: object
        lang:
          type: string
        language:
          type: string
        lastUpdate:
          type: string
          format: date-time
        score:
          type: number
        status:
          type: integer
        title:
          type: string
        translations:
          type: array
          items:
            type: object
    Materials:
      type: array
      items:
        $ref: '#/components/schemas/Material'
    GetMaterialResponse:
      properties:
        material:
          type: object
...